test(net-stubbing): consolidate core specs and update wiring docs
Split duplicated core coverage into focused specs, fold driver-events into
cy-intercept, and refresh README wiring for CyIntercept.
