+++ packages/net-stubbing/test/unit/core/merge-handler-result.spec.ts
@@ -1,105 +1,8 @@
-import { describe, it, expect, vi } from 'vitest'
-import {
-  doesRouteMatch,
-  getMatchableForRequest,
-  matchRoutes,
-} from '../../../lib/core/route-matching'
-import { planSubscriptions } from '../../../lib/core/plan-subscriptions'
+import { describe, it, expect } from 'vitest'
 import {
   mergeIncomingRequestChanges,
   mergeIncomingResponseChanges,
 } from '../../../lib/core/merge-handler-result'
-import type { BackendRoute, RouteMatcherOptions } from '@packages/network-interception'
-
-describe('core/route-matching', () => {
-  const tryMatch = (req: { url: string, method?: string, headers?: Record<string, string> }, matcher: RouteMatcherOptions, expected = true) => {
-    expect(doesRouteMatch(matcher, {
-      method: 'GET',
-      headers: {},
-      ...req,
-    })).toEqual(expected)
-  }
-
-  it('matches exact URL', () => {
-    tryMatch({ url: 'https://google.com/foo' }, { url: 'https://google.com/foo' })
-  })
-
-  it('matches globs against path', () => {
-    tryMatch({ url: 'http://foo.com/bar/a1' }, { url: '/bar/*' })
-  })
-
-  it('orders middleware routes before handlers', () => {
-    const routes = [
-      { id: '1', routeMatcher: { middleware: true, pathname: '/foo' }, hasInterceptor: false, getFixture: async () => {}, matches: 0 },
-      { id: '2', routeMatcher: { pathname: '/foo' }, hasInterceptor: false, getFixture: async () => {}, matches: 0 },
-      { id: '3', routeMatcher: { middleware: true, pathname: '/foo' }, hasInterceptor: false, getFixture: async () => {}, matches: 0 },
-      { id: '4', routeMatcher: { pathname: '/foo' }, hasInterceptor: false, getFixture: async () => {}, matches: 0 },
-    ] as BackendRoute[]
-
-    const matched = matchRoutes(routes, {
-      method: 'GET',
-      headers: {},
-      url: 'http://bar.baz/foo?_',
-    })
-
-    expect(matched.map((r) => r.id)).toEqual(['1', '3', '4', '2'])
-  })
-
-  it('getMatchableForRequest extracts auth from basic header', () => {
-    const matchable = getMatchableForRequest({
-      headers: { authorization: 'basic Zm9vOmJhcg==' },
-      method: 'GET',
-      url: 'https://google.com/asdf?1234=a',
-    })
-
-    expect(matchable.auth).toEqual({ username: 'foo', password: 'bar' })
-  })
-})
-
-describe('core/plan-subscriptions', () => {
-  it('plans default subscriptions for matched routes', () => {
-    const routes = [{
-      id: 'route-1',
-      hasInterceptor: true,
-      routeMatcher: {},
-      getFixture: async () => {},
-      matches: 0,
-    }] as BackendRoute[]
-
-    const planned = planSubscriptions({ matchingRoutes: routes })
-
-    expect(planned).toHaveLength(1)
-    expect(planned[0].subscriptions.map((s) => s.eventName)).toEqual([
-      'before:request',
-      'response:callback',
-      'after:response',
-      'network:error',
-    ])
-
-    expect(planned[0].subscriptions[0].await).toBe(true)
-  })
-
-  it('skips sync XHR routes with interceptors', () => {
-    const onSyncInterceptSkipped = vi.fn()
-    const routes = [{
-      id: 'route-1',
-      hasInterceptor: true,
-      routeMatcher: {},
-      getFixture: async () => {},
-      matches: 0,
-    }] as BackendRoute[]
-
-    const planned = planSubscriptions({
-      matchingRoutes: routes,
-      isSyncRequest: true,
-      url: 'http://example.com',
-      onSyncInterceptSkipped,
-    })
-
-    expect(planned).toHaveLength(0)
-    expect(onSyncInterceptSkipped).toHaveBeenCalledWith('http://example.com')
-  })
-})
 
 describe('core/merge-handler-result', () => {
   it('merges handler changes and resolves relative URLs', () => {
