+++ packages/network-interception/README.md
@@ -20,7 +20,7 @@ The codebase adopts **hexagonal architecture** (also called **ports and adapters
 | --- | --- | --- |
 | **Port** | Contract at the edge of the interception "inside" | `For*` types in `lib/ports/` |
 | **Adapter** | Implements a port by delegating to existing Cypress code | `*Adapter` classes under `packages/*/lib/adapters/` |
-| **Driving port** (primary) | Outside actors **call into** interception | `ForHttpIntercept`, `ForInterceptRegistration` |
+| **Driving port** (primary) | Outside actors **call into** interception | `ForHttpIntercept` |
 | **Driven port** (secondary) | Interception **calls out** for I/O | `ForInterceptionEvents`, `ForCookieState`, … |
 | **Core** | Domain orchestration without transport imports | **`HttpIntercept`** |
 | **Composition root** | Constructs and injects adapters + core | `server-base.open()` |
@@ -60,9 +60,9 @@ type InterceptMiddleware = (
 flowchart TB
   subgraph compositionRoot["server-base open"]
     REG["registerDefaultInterceptMiddleware<br/>blockHosts + CSP"]
-    DRV["createDriverAdapter<br/>cy.intercept intercepter"]
+    CI["CyIntercept<br/>cy.intercept middleware"]
     REG --> HI[HttpIntercept]
-    DRV --> HI
+    CI --> HI
   end
   subgraph transportLayer["transport adapters"]
     MITM["ApplyHttpInterception<br/>packages/proxy"]
@@ -79,8 +79,8 @@ flowchart TB
 ```
 server-base.open()
   → createHttpInterceptWithDefaultMiddleware() (blockHosts + CSP middleware)
-  → createDriverAdapter() (registers cy.intercept intercepter via use())
-  → if proxy enabled: createNetworkProxy() with Proxy*Adapter driven ports
+  → new CyIntercept() + httpIntercept.use(cyIntercept.middleware)
+  → if proxy enabled: createNetworkProxy() with proxy driven ports
   → if proxy disabled: open_project passes same HttpIntercept to CDPNetworkInterception
 
 ApplyHttpInterception middleware (proxy path)
@@ -104,7 +104,7 @@ Use this when a concern can run on materialized `HttpRequest` / `HttpResponse` (
 | --- | --- | --- | --- |
 | `blockHosts` | `createBlockedHostsInterceptMiddleware` (`network-interception`) | `registerDefaultInterceptMiddleware` | Yes (same stack) |
 | CSP allow-list | `createCspAllowListInterceptMiddleware` (`network-interception`) | `registerDefaultInterceptMiddleware` | Yes (same stack) |
-| `cy.intercept` | `CyInterceptIntercepter` (`net-stubbing`) | `server-base.open()` via `createDriverAdapter` | Yes |
+| `cy.intercept` | `CyIntercept` (`net-stubbing`) | `server-base.open()` via `httpIntercept.use()` | Yes |
 | Document rewrite | `SetInjectionLevel` → `MaybeInjectHtml` → `MaybeRemoveSecurity` (`proxy` streaming) | `server-base` wires `createProxyNetworkServices()` driven ports | No (intentional) |
 
 **Keep on proxy streaming middleware** when the concern needs the response byte stream (gzip/br), multiple response-stage hooks, or is proxy-only for now:
@@ -137,7 +137,7 @@ Document rewrite (`modifyObstructiveCode`, `experimentalModifyObstructiveThirdPa
 
 System-level `cy.intercept` smoke with proxy disabled: `system-tests/test/cdp_network_intercept_spec.js` (`fetch.cy.js` under `CYPRESS_INTERNAL_DISABLE_PROXY=1`).
 
-**Tests:** `@packages/browser-automation` unit specs use a mocked `handle` (CDP transport only). `cy.intercept` stack integration (real `HttpIntercept` + driver adapter) lives in `@packages/server` (`test/unit/browsers/cdp-cy-intercept-integration_spec.ts`).
+**Tests:** `@packages/browser-automation` unit specs use a mocked `handle` (CDP transport only). `cy.intercept` stack integration (real `HttpIntercept` + `CyIntercept`) lives in `@packages/server` (`test/unit/browsers/cdp-cy-intercept-integration_spec.ts`).
 
 ---
